Reorganize cfgs to reduce `#[cfg]` quantities
